requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cos-python-sdk-v5==1.8.1
dicttoxml==1.7.4
Django==3.0.8
django-haystack==2.8.1
idna==2.10
jieba==0.42.1
mysqlclient==2.0.1
pytz==2020.1
qcloud-python-sts==3.0.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
Whoosh==2.7.4
